UI : Move crud service interface in admin directory
